requirements-prod.txt
asgiref==3.3.1
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
Django==3.0.3
django-cors-headers==3.7.0
django-filter==2.4.0
django-ranged-response==0.2.0
django-xff==1.3.0
djangorestframework==3.11.0
django-rest-framework-proxy==1.6.0
django-utils==0.0.2
django-utils-six==2.0
django-simple-captcha==0.5.14
django-import-export==2.5.0
idna==2.10
lxml==4.6.2
mysqlclient
python-docx==0.8.10
requests==2.25.1
six==1.15.0
urllib3==1.26.2
xlwt==1.3.0
uwsgi==2.0.19.1
ddt
faker
pycryptodomex
https://huoqi-public.oss-cn-beijing.aliyuncs.com/iast/dongtai-latest.tar.gz
